requirements.txt
diffusers==0.37.0
huggingface_hub==1.7.1
joblib==1.5.3
numpy==1.26.4
ollama==0.6.1
Pillow==12.2.0
pydantic==2.13.0
Requests==2.33.1
sentence_transformers==5.3.0
soundfile==0.13.1
spacy==3.8.11
streamlit==1.56.0
style_bert_vits2==2.5.0
SudachiPy==0.6.10
tqdm==4.67.3
transformers==5.3.0
SudachiDict-core==20260116
SudachiDict-full==20260116
ja-ginza==5.2.0
accelerate==1.13.0
peft==0.18.1
xgboost==3.2.0